pernod_ricard_agent_repo_full/requirements.txt
streamlit
requests
httpx
beautifulsoup4
readability-lxml
psycopg2-binary
sqlalchemy
pydantic
python-dotenv
openai
tqdm
python-dateutil
pgvector
pytest
